% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Scales a numeric vector using min_max scaling method}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A scaled numeric vector
}
\description{
Scales a numeric vector using min_max scaling method
}
\examples{
x <- 1:5
min_max_scale(x)
}
