% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Sets theme of your ggplot2 plot to an academic acceptable theme}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{Desired font size, default is 12}
}
\value{
A plot with the adjusted theme elements
}
\description{
Sets theme of your ggplot2 plot to an academic acceptable theme
}
